Added Print Functionality
Used JavaScript to activate the Print This Page button.  Clicking on the button redirects the user to a printer friendly version of the page which does not contain the "Print" button.  I also added an @media CSS directive in site.css to eliminate the URL header and footer when printing.  Not sure if this will work on all browsers, but it does work on Chrome.