Completed Official Solution
Typed-along with Bobby as he explained his solution.  He placed the print button in a d-flex div instead of making it a row and used the d-print-none bs class to make the button dissapear on print, which is more efficient than my "duplicate" page strategy.  Also, he pointed out that "row" is already a d-flex, which allows us to use align & justify classes.   He used table-striped for table formatting and included a <tfooter> to "reset" the striped behavior.  He also made i more responsive by adding a screen media query to the CSS to make the font smaller for some elements.  In CSS, we can effect changes based on class, id, or element.  Advice:  Try to avoid element level.